# [perf] Use local schemas if available