… tests/test_issues/input/issue_1040.yaml