@cmungall 2024-03-18
Thanks! Yes, we should be phasing out the older hbread code over time